+++ tests/test_cmp.py
@@ -2,7 +2,9 @@
 
 import subprocess
 import sys
-from ppu_tu import setup, teardown, find_in_path  # noqa
+
+from ppu_tu import find_in_path
+from ppu_tu import tmp_dir  # noqa: F401 tmp_dir imported but unused
 
 
 test_prog_path = find_in_path('cmp.py')