+++ tests/test_remove_old_files.py
@@ -4,8 +4,10 @@ from time import time
 import os
 import subprocess
 import sys
-from ppu_tu import setup, teardown, find_in_path  # noqa
+
 from ppu_tu import create_files, assert_files_exist, assert_files_not_exist
+from ppu_tu import find_in_path
+from ppu_tu import tmp_dir  # noqa: F401 tmp_dir imported but unused
 
 
 test_prog_path = find_in_path('remove-old-files.py')
@@ -37,11 +39,6 @@ def test_recursive():
 
 
 def test_remove_empty_directory():
-    teardown()
-    try:
-        setup()
-    except OSError:
-        pass
     create_files(['test3', 'test4'], 'subdir')
     test3 = os.path.join('subdir', 'test3')
     test4 = os.path.join('subdir', 'test4')